api/requirements.txt
Django
djangorestframework
django-cors-headers
requests
python-dotenv
beautifulsoup4
selenium
webdriver-manager
